+++ src/store/modules/navigation.js
@@ -110,22 +110,35 @@ const navigation = {
   toView (options) {
     if (typeof options === 'string') {
       store.actions.setView(options)
-    } else {
-      if (!options.to) store.actions.setLog('Programmatic navigation: "to" is required ', 'error')
-      if (!options.fromSpot) store.actions.setLog('Programmatic navigation: "fromSpot" is required ', 'error')
-      if (options.fromSpot && typeof options.fromSpot !== 'object') store.actions.setLog('Programmatic navigation: "fromSpot" should be an object ', 'error')
-      if (options.params && typeof options.params !== 'object') store.actions.setLog('Programmatic navigation: "params" should be an object ', 'error')
-      if (options.to && options.fromSpot) {
-        if (!options.params) options.params = {}
-        const positionParams = options.fromSpot.position ? { position: { X: options.fromSpot.position.Xabs, Y: options.fromSpot.position.Yabs, scale: options.fromSpot.position.scale, Xi: options.fromSpot.position.Xi, Yi: options.fromSpot.position.Yi, scalei: options.fromSpot.position.scalei } } : {}
-        store.actions.setView(
-          {
-            name: options.to,
-            params: options.params
-          },
-          positionParams
-        )
+      return
+    }
+    if (!options.to) store.actions.setLog('Programmatic navigation: "to" is required ', 'error')
+    if (!options.fromSpot) store.actions.setLog('Programmatic navigation: "fromSpot" is required ', 'error')
+    if (options.fromSpot && typeof options.fromSpot !== 'object') store.actions.setLog('Programmatic navigation: "fromSpot" should be an object ', 'error')
+    if (options.params && typeof options.params !== 'object') store.actions.setLog('Programmatic navigation: "params" should be an object ', 'error')
+    if (options.to && options.fromSpot) {
+      if (!options.params) {
+        options.params = {}
       }
+      const positionParams = options.fromSpot.position
+        ? {
+            position: {
+              X: options.fromSpot.position.Xabs,
+              Y: options.fromSpot.position.Yabs,
+              scale: options.fromSpot.position.scale,
+              Xi: options.fromSpot.position.Xi,
+              Yi: options.fromSpot.position.Yi,
+              scalei: options.fromSpot.position.scalei
+            }
+          }
+        : {}
+      store.actions.setView(
+        {
+          name: options.to,
+          params: options.params
+        },
+        positionParams
+      )
     }
   },
   setView (data, options) {
@@ -136,7 +149,7 @@ const navigation = {
 
     const view = parseView(data)
     const position = (
-      !options || (options.position && options.position.scale === 0)
+      !options || !options.position || options.position.scale === 0
         ? {
             X: 0,
             Y: 0,