+++ src/store/modules/position.js
@@ -1,4 +1,5 @@
 import store from '../store'
+
 function calcCoords (distance, angle, parentSize) {
   let X = 0
   let Y = 0
@@ -7,8 +8,8 @@ function calcCoords (distance, angle, parentSize) {
   if (distance !== 0) {
     X = ((store.actions.getComponentWidth(parentSize) / 2) * distance / 100) * Math.cos(angle * (Math.PI / 180))
     Y = ((store.actions.getComponentWidth(parentSize) / 2) * distance / 100) * Math.sin(angle * (Math.PI / 180))
-    Xi = X > 0 ? -Math.abs(Number(X)) : Math.abs(Number(X))
-    Yi = Y > 0 ? -Math.abs(Number(Y)) : Math.abs(Number(Y))
+    Xi = -X
+    Yi = -Y
   }
   return {
     X,
@@ -17,6 +18,7 @@ function calcCoords (distance, angle, parentSize) {
     Yi
   }
 }
+
 function determinePosition (pos) {
   if (store.state.history[store.state.history.length - pos]) {
     return store.state.history[store.state.history.length - pos].position
@@ -26,9 +28,10 @@ function determinePosition (pos) {
     }
   }
 }
+
 const position = {
   getCurrentPosition () {
-    return store.state.history[store.state.history.length - 1].position
+    return determinePosition(1)
   },
   getPreviousPosition () {
     return determinePosition(2)
@@ -41,11 +44,18 @@ const position = {
     return store.actions.getCurrentViewName() === viewName ? store.actions.getCurrentPosition() : store.actions.getPastPosition()
   },
   calcPosition (component) {
-    // store.actions.setLog('calcPosition() => ' + component.componentType)
-    // Variable declaration
-    let parentPosition = { Xi: 0, Yi: 0, X: 0, Y: 0, scalei: 1, scale: 1 }
+    const parentPosition =
+      ['z-view', 'z-list', 'z-spot'].includes(component.$parent.componentType)
+        ? {
+            Xi: component.$parent.position.Xi,
+            Yi: component.$parent.position.Yi,
+            X: component.$parent.position.X,
+            Y: component.$parent.position.Y,
+            scalei: component.$parent.position.scalei,
+            scale: component.$parent.position.scale
+          }
+        : { Xi: 0, Yi: 0, X: 0, Y: 0, scalei: 1, scale: 1 }
     const newCoords = calcCoords(component.distance, component.angle, component.$parent.size)
-    if (component.$parent.componentType === 'z-view' || component.$parent.componentType === 'z-list' || component.$parent.componentType === 'z-spot') parentPosition = { Xi: component.$parent.position.Xi, Yi: component.$parent.position.Yi, X: component.$parent.position.X, Y: component.$parent.position.Y, scalei: component.$parent.position.scalei, scale: component.$parent.position.scale }
     return {
       X: newCoords.X,
       Y: newCoords.Y,